fs.existsSync instead of File.exists
I'm reading this code and I wonder howit worked before 🤔  but now that I upgraded Laravel Mix and PostCSS in my dependencies, there's a node error on "File.exists" and says "File is not defined". Might be a trailing global accessible in previous versions of node? Or something that webpack passed before but doesn't anymore?